Refactor connect_page getProfile function
