+++ lib/model/sensor_model/sensor.dart
