+++ lib/screens/auth/forget_page.dart
@@ -3,26 +3,21 @@ import 'package:provider/provider.dart';
 import 'package:smartfarm/animation/fade_animation.dart';
 import 'package:smartfarm/firebase/auth_exception_handler.dart';
 import 'package:smartfarm/firebase/auth_result_status.dart';
-import 'package:smartfarm/provider/firebase_provider.dart';
-import 'package:smartfarm/screen/forget_page.dart';
-import 'package:smartfarm/screen/signup_page.dart';
-import 'package:smartfarm/shared/smartfarmer_constants.dart';
+import 'package:smartfarm/services/firebase_provider.dart';
 
-class LoginPage extends StatefulWidget {
+class ForgetPage extends StatefulWidget {
   @override
-  _LoginPageState createState() => _LoginPageState();
+  _ForgetPageState createState() => _ForgetPageState();
 }
 
-class _LoginPageState extends State<LoginPage> {
+class _ForgetPageState extends State<ForgetPage> {
   TextEditingController _mailCon = TextEditingController();
-  TextEditingController _pwCon = TextEditingController();
   final GlobalKey<ScaffoldState> _scaffoldKey = new GlobalKey<ScaffoldState>();
   FirebaseProvider fp;
 
   @override
   void dispose() {
     _mailCon.dispose();
-    _pwCon.dispose();
     super.dispose();
   }
 
@@ -50,10 +45,9 @@ class _LoginPageState extends State<LoginPage> {
       ),
       body: SingleChildScrollView(
         child: Container(
-          height: MediaQuery.of(context).size.height,
+          height: MediaQuery.of(context).size.height * 0.5,
           width: double.infinity,
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
               Expanded(
                 child: Column(
@@ -61,8 +55,10 @@ class _LoginPageState extends State<LoginPage> {
                   children: <Widget>[
                     Column(
                       children: <Widget>[
-                        FadeAnimation(1, Text(
-                            "로그인",
+                        FadeAnimation(
+                          1,
+                          Text(
+                            "비밀번호 찾기",
                             style: TextStyle(
                               fontFamily: 'NotoSans-Bold',
                               fontSize: 32.0,
@@ -72,8 +68,10 @@ class _LoginPageState extends State<LoginPage> {
                         SizedBox(
                           height: 20,
                         ),
-                        FadeAnimation(1.2, Text(
-                            "이메일과 비밀번호를 입력해주세요",
+                        FadeAnimation(
+                          1.2,
+                          Text(
+                            "가입했던 이메일을 입력해주세요",
                             textAlign: TextAlign.center,
                             style: TextStyle(
                               color: Colors.grey[700],
@@ -88,32 +86,25 @@ class _LoginPageState extends State<LoginPage> {
                       padding: EdgeInsets.symmetric(horizontal: 40),
                       child: Column(
                         children: <Widget>[
-                          FadeAnimation(1.2, makeInput(label: '이메일', editingController: _mailCon),),
+                          FadeAnimation(
+                            1.2,
+                            makeInput(
+                                label: '이메일', editingController: _mailCon),
+                          ),
                           SizedBox(
                             height: 30,
                           ),
-                          FadeAnimation(1.3, makeInput(label: '비밀번호', obscureText: true, editingController: _pwCon)),
                           SizedBox(
                             height: 6,
-                          ),
-                          Align(
-                            alignment: Alignment.centerRight,
-                            child: InkWell(
-                              onTap: (){
-                                Navigator.push(context, MaterialPageRoute(builder: (context) => ForgetPage()));
-                              },
-                              child: FadeAnimation(1.3, Text("비밀번호 찾기", style: TextStyle(
-                                  fontFamily: 'NotoSans-Medium'
-                                ),),
-                              ),
-                            ),
                           ),
                         ],
                       ),
                     ),
                     Padding(
                       padding: EdgeInsets.symmetric(horizontal: 40),
-                      child: FadeAnimation(1.4, Container(
+                      child: FadeAnimation(
+                        1.4,
+                        Container(
                           padding: EdgeInsets.only(top: 3, left: 3),
                           decoration: BoxDecoration(
                             borderRadius: BorderRadius.circular(50),
@@ -128,16 +119,17 @@ class _LoginPageState extends State<LoginPage> {
                             minWidth: double.infinity,
                             height: 60,
                             onPressed: () {
-                              FocusScope.of(context).requestFocus(new FocusNode());
-                              _logIn();
+                              FocusScope.of(context)
+                                  .requestFocus(new FocusNode());
+                              _findPassword();
                             },
-                            color: Colors.greenAccent,
+                            color: Colors.purpleAccent,
                             elevation: 0,
                             shape: RoundedRectangleBorder(
                               borderRadius: BorderRadius.circular(50),
                             ),
                             child: Text(
-                              '로그인',
+                              '전송',
                               style: TextStyle(
                                 fontFamily: 'NotoSans-Medium',
                                 fontSize: 18.0,
@@ -147,39 +139,9 @@ class _LoginPageState extends State<LoginPage> {
                         ),
                       ),
                     ),
-                    FadeAnimation(1.5, Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: <Widget>[
-                          Text(
-                            "계정이 없으신가요? ",
-                            style: TextStyle(fontFamily: 'Notosans-Regular'),
-                          ),
-                          InkWell(
-                            onTap: (){
-                              Navigator.push(context, MaterialPageRoute(builder: (context) => SignUpPage()));
-                            },
-                            child: Text(
-                              "가입하기",
-                              style: TextStyle(
-                                fontFamily: 'Notosans-Bold',
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
                   ],
                 ),
               ),
-              FadeAnimation(1.2, Container(
-                  height: MediaQuery.of(context).size.height / 3,
-                  decoration: BoxDecoration(
-                    image: DecorationImage(
-                      image: AssetImage(farmerWomen),
-                    ),
-                  ),
-                ),
-              )
             ],
           ),
         ),
@@ -187,7 +149,7 @@ class _LoginPageState extends State<LoginPage> {
     );
   }
 
-  _logIn() async{
+  _findPassword() async {
     _scaffoldKey.currentState
       ..hideCurrentSnackBar()
       ..showSnackBar(SnackBar(
@@ -195,18 +157,21 @@ class _LoginPageState extends State<LoginPage> {
         content: Row(
           children: <Widget>[
             CircularProgressIndicator(),
-            Text("   로그인 중입니다...")
+            Text("   이메일 전송중입니다...")
           ],
         ),
       ));
-    final result = await fp.signInWithEmail(_mailCon.text, _pwCon.text);
+
+    final result = await fp.sendPasswordResetEmail(_mailCon.text);
+    print('result 상태 $result');
     _scaffoldKey.currentState.hideCurrentSnackBar();
-    if (result == AuthResultStatus.successful){
-      Navigator.pop(context);
-    }else{
+    if (result == AuthResultStatus.successful) {
+      _mailCon.text = '';
+      showMessage('메일을 전송했습니다. 이메일을 확인해주세요.');
+    } else {
       final errorMsg = AuthExceptionHandler.generateExceptionMessage(result);
       showLastFBMessage(errorMsg);
-      _pwCon.text = '';
+      _mailCon.text = '';
     }
   }
 
@@ -224,9 +189,23 @@ class _LoginPageState extends State<LoginPage> {
         ),
       ));
   }
+
+  showMessage(String msg) {
+    _scaffoldKey.currentState
+      ..hideCurrentSnackBar()
+      ..showSnackBar(SnackBar(
+        duration: Duration(seconds: 10),
+        content: Text(msg),
+        action: SnackBarAction(
+          label: "Done",
+          onPressed: () {},
+        ),
+      ));
+  }
 }
 
-Widget makeInput({label, obscureText = false, TextEditingController editingController}) {
+Widget makeInput(
+    {label, obscureText = false, TextEditingController editingController}) {
   return Column(
     crossAxisAlignment: CrossAxisAlignment.start,
     children: <Widget>[
