+++ lib/screens/auth/login_page.dart
@@ -3,23 +3,27 @@ import 'package:provider/provider.dart';
 import 'package:smartfarm/animation/fade_animation.dart';
 import 'package:smartfarm/firebase/auth_exception_handler.dart';
 import 'package:smartfarm/firebase/auth_result_status.dart';
-import 'package:smartfarm/provider/firebase_provider.dart';
-import 'package:smartfarm/screen/signup_page.dart';
+import 'package:smartfarm/screens/auth/forget_page.dart';
+import 'package:smartfarm/screens/auth/signup_page.dart';
+import 'package:smartfarm/services/firebase_provider.dart';
+
 import 'package:smartfarm/shared/smartfarmer_constants.dart';
 
-class ForgetPage extends StatefulWidget {
+class LoginPage extends StatefulWidget {
   @override
-  _ForgetPageState createState() => _ForgetPageState();
+  _LoginPageState createState() => _LoginPageState();
 }
 
-class _ForgetPageState extends State<ForgetPage> {
+class _LoginPageState extends State<LoginPage> {
   TextEditingController _mailCon = TextEditingController();
+  TextEditingController _pwCon = TextEditingController();
   final GlobalKey<ScaffoldState> _scaffoldKey = new GlobalKey<ScaffoldState>();
   FirebaseProvider fp;
 
   @override
   void dispose() {
     _mailCon.dispose();
+    _pwCon.dispose();
     super.dispose();
   }
 
@@ -47,9 +51,10 @@ class _ForgetPageState extends State<ForgetPage> {
       ),
       body: SingleChildScrollView(
         child: Container(
-          height: MediaQuery.of(context).size.height * 0.5,
+          height: MediaQuery.of(context).size.height,
           width: double.infinity,
           child: Column(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
               Expanded(
                 child: Column(
@@ -57,10 +62,8 @@ class _ForgetPageState extends State<ForgetPage> {
                   children: <Widget>[
                     Column(
                       children: <Widget>[
-                        FadeAnimation(
-                          1,
-                          Text(
-                            "비밀번호 찾기",
+                        FadeAnimation(1, Text(
+                            "로그인",
                             style: TextStyle(
                               fontFamily: 'NotoSans-Bold',
                               fontSize: 32.0,
@@ -70,10 +73,8 @@ class _ForgetPageState extends State<ForgetPage> {
                         SizedBox(
                           height: 20,
                         ),
-                        FadeAnimation(
-                          1.2,
-                          Text(
-                            "가입했던 이메일을 입력해주세요",
+                        FadeAnimation(1.2, Text(
+                            "이메일과 비밀번호를 입력해주세요",
                             textAlign: TextAlign.center,
                             style: TextStyle(
                               color: Colors.grey[700],
@@ -88,25 +89,32 @@ class _ForgetPageState extends State<ForgetPage> {
                       padding: EdgeInsets.symmetric(horizontal: 40),
                       child: Column(
                         children: <Widget>[
-                          FadeAnimation(
-                            1.2,
-                            makeInput(
-                                label: '이메일', editingController: _mailCon),
-                          ),
+                          FadeAnimation(1.2, makeInput(label: '이메일', editingController: _mailCon),),
                           SizedBox(
                             height: 30,
                           ),
+                          FadeAnimation(1.3, makeInput(label: '비밀번호', obscureText: true, editingController: _pwCon)),
                           SizedBox(
                             height: 6,
+                          ),
+                          Align(
+                            alignment: Alignment.centerRight,
+                            child: InkWell(
+                              onTap: (){
+                                Navigator.push(context, MaterialPageRoute(builder: (context) => ForgetPage()));
+                              },
+                              child: FadeAnimation(1.3, Text("비밀번호 찾기", style: TextStyle(
+                                  fontFamily: 'NotoSans-Medium'
+                                ),),
+                              ),
+                            ),
                           ),
                         ],
                       ),
                     ),
                     Padding(
                       padding: EdgeInsets.symmetric(horizontal: 40),
-                      child: FadeAnimation(
-                        1.4,
-                        Container(
+                      child: FadeAnimation(1.4, Container(
                           padding: EdgeInsets.only(top: 3, left: 3),
                           decoration: BoxDecoration(
                             borderRadius: BorderRadius.circular(50),
@@ -121,17 +129,16 @@ class _ForgetPageState extends State<ForgetPage> {
                             minWidth: double.infinity,
                             height: 60,
                             onPressed: () {
-                              FocusScope.of(context)
-                                  .requestFocus(new FocusNode());
-                              _findPassword();
+                              FocusScope.of(context).requestFocus(new FocusNode());
+                              _logIn();
                             },
-                            color: Colors.purpleAccent,
+                            color: Colors.greenAccent,
                             elevation: 0,
                             shape: RoundedRectangleBorder(
                               borderRadius: BorderRadius.circular(50),
                             ),
                             child: Text(
-                              '전송',
+                              '로그인',
                               style: TextStyle(
                                 fontFamily: 'NotoSans-Medium',
                                 fontSize: 18.0,
@@ -141,9 +148,39 @@ class _ForgetPageState extends State<ForgetPage> {
                         ),
                       ),
                     ),
+                    FadeAnimation(1.5, Row(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: <Widget>[
+                          Text(
+                            "계정이 없으신가요? ",
+                            style: TextStyle(fontFamily: 'Notosans-Regular'),
+                          ),
+                          InkWell(
+                            onTap: (){
+                              Navigator.push(context, MaterialPageRoute(builder: (context) => SignUpPage()));
+                            },
+                            child: Text(
+                              "가입하기",
+                              style: TextStyle(
+                                fontFamily: 'Notosans-Bold',
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
                   ],
                 ),
               ),
+              FadeAnimation(1.2, Container(
+                  height: MediaQuery.of(context).size.height / 3,
+                  decoration: BoxDecoration(
+                    image: DecorationImage(
+                      image: AssetImage(farmerWomen),
+                    ),
+                  ),
+                ),
+              )
             ],
           ),
         ),
@@ -151,7 +188,7 @@ class _ForgetPageState extends State<ForgetPage> {
     );
   }
 
-  _findPassword() async {
+  _logIn() async{
     _scaffoldKey.currentState
       ..hideCurrentSnackBar()
       ..showSnackBar(SnackBar(
@@ -159,21 +196,18 @@ class _ForgetPageState extends State<ForgetPage> {
         content: Row(
           children: <Widget>[
             CircularProgressIndicator(),
-            Text("   이메일 전송중입니다...")
+            Text("   로그인 중입니다...")
           ],
         ),
       ));
-
-    final result = await fp.sendPasswordResetEmail(_mailCon.text);
-    print('result 상태 $result');
+    final result = await fp.signInWithEmail(_mailCon.text, _pwCon.text);
     _scaffoldKey.currentState.hideCurrentSnackBar();
-    if (result == AuthResultStatus.successful) {
-      _mailCon.text = '';
-      showMessage('메일을 전송했습니다. 이메일을 확인해주세요.');
-    } else {
+    if (result == AuthResultStatus.successful){
+      Navigator.pop(context);
+    }else{
       final errorMsg = AuthExceptionHandler.generateExceptionMessage(result);
       showLastFBMessage(errorMsg);
-      _mailCon.text = '';
+      _pwCon.text = '';
     }
   }
 
@@ -191,23 +225,9 @@ class _ForgetPageState extends State<ForgetPage> {
         ),
       ));
   }
-
-  showMessage(String msg) {
-    _scaffoldKey.currentState
-      ..hideCurrentSnackBar()
-      ..showSnackBar(SnackBar(
-        duration: Duration(seconds: 10),
-        content: Text(msg),
-        action: SnackBarAction(
-          label: "Done",
-          onPressed: () {},
-        ),
-      ));
-  }
 }
 
-Widget makeInput(
-    {label, obscureText = false, TextEditingController editingController}) {
+Widget makeInput({label, obscureText = false, TextEditingController editingController}) {
   return Column(
     crossAxisAlignment: CrossAxisAlignment.start,
     children: <Widget>[
