+++ lib/screens/farm_list/farm_list_page.dart
@@ -2,12 +2,13 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter_swiper/flutter_swiper.dart';
 import 'package:provider/provider.dart';
-import 'package:smartfarm/model/profile_farmer.dart';
+import 'package:smartfarm/model/farmer_model/profile_farmer.dart';
+
 import 'package:http/http.dart' as http;
-import 'package:smartfarm/provider/firebase_provider.dart';
-import 'package:smartfarm/provider/scan_data.dart';
-import 'package:smartfarm/screen/connect_page.dart';
-import 'package:smartfarm/screen/info_page.dart';
+import 'package:smartfarm/services/firebase_provider.dart';
+import 'package:smartfarm/services/scan_data.dart';
+import 'package:smartfarm/screens/devices_connect/connect_page.dart';
+import 'package:smartfarm/screens/farm_dashboard/info_page.dart';
 import 'package:smartfarm/shared/smartfarmer_constants.dart';
 
 class FarmListPage extends StatelessWidget {
@@ -116,10 +117,7 @@ class FarmListPage extends StatelessWidget {
                                                               context,
                                                               listen: false)
                                                           .setDeviceUUID(
-                                                              farmerProfile
-                                                                  .farmInfo[
-                                                                      index]
-                                                                  .deviceUUID);
+                                                              farmerProfile.farmInfo[index].deviceUUID);
                                                       Navigator.push(
                                                         context,
                                                         MaterialPageRoute(
