+++ lib/services/scan_data.dart
