+++ lib/utils/transfomer.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:smartfarm/model/farmer.dart';
+import 'package:smartfarm/model/farmer_model/farmer.dart';
 
 class Transfomer {
   final transFarmerData = StreamTransformer<DocumentSnapshot, Farmer>.fromHandlers(
